+++ test/fixtures/test.php
@@ -1,0 +1,10 @@
+<!doctype html>
+<html lang="en">
+<head>
+    <meta charset="utf-8">
+    <title>test</title>
+</head>
+<body>
+    <?php echo '<p>test</p>'; ?>
+</body>
+</html>