+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### HEAD
 
+* Add configs that remove the `X-Powered-By` HTTP response header
+  [[#54](https://github.com/h5bp/server-configs-apache/issues/54)].
 * Add expires rules for WebP
   [[#61](https://github.com/h5bp/server-configs-apache/issues/61)].